reorganizing client to have subparsers for better organization
