+++ setup.py
@@ -94,5 +94,5 @@ if __name__ == "__main__":
               'Programming Language :: Python :: 2.7',
               'Programming Language :: Python :: 3',
           ],
-          entry_points = {'console_scripts': [ 'shub=singularity.scripts:main',
+          entry_points = {'console_scripts': [ 'shub=singularity.cli:main',
                                                'sreg=singularity.registry.client:main' ] })
