+++ singularity/cli/__init__.py
@@ -2,7 +2,7 @@
 
 The MIT License (MIT)
 
-Copyright (c) 2016-2017 Vanessa Sochat
+Copyright (c) 2016-2017 Vanessa Sochat, All Rights Reserved
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -24,8 +24,6 @@ SOFTWARE.
 
 '''
 
-from singularity.version import __version__
-
-__all__ = ['analysis', 'build', 'views', 'api',
-           'app', 'cli', 'logger', 'registry', 
-           'utils', '__version__']
+from .wrapper import Singularity
+from .base import main
+from .utils import get_image
