+++ singularity/cli/base/package.py
@@ -1,0 +1,66 @@
+'''
+
+Copyright (c) 2016-2017 Vanessa Sochat, All Rights Reserved
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+
+'''
+
+from singularity.utils import check_install
+from singularity.package import package
+from singularity.logger import bot
+from singularity.cli import get_image
+import sys
+import os
+
+
+def main(args,parser,subparser):
+
+    # We can only continue if singularity is installed
+    if check_install() is not True:
+        bot.error("Cannot find Singularity! Is it installed?")
+        sys.exit(1)
+
+    # Output folder will be pwd if not specified
+    output_folder = os.getcwd()
+    if args.outfolder is not None:
+        output_folder = os.getcwd()
+
+
+    # If we are given an image, ensure full path
+    if args.image is not None:
+
+        image,existed = get_image(args.image,
+                                  return_existed=True,
+                                  size=args.size)
+
+        if image is None:
+            bot.error("Cannot find image. Exiting.")
+            sys.exit(1)
+
+        remove_image = not args.include_image
+        package(image_path=image,
+                output_folder=output_folder,
+                runscript=True,
+                software=True,
+                remove_image=remove_image)
+
+    else:
+        print("Please specify an image to package with --image")
+        subparser.print_help()
