+++ singularity/cli/static/img/folder-small.png
