+++ singularity/cli/templates/container_tree.html
