+++ singularity/cli/templates/container_tree_circleci.html
