+++ singularity/cli/templates/index_circleci.html
