+++ singularity/cli/templates/similarity_scatter.html
